rapid_deepfake_mvp/requirements.txt
torch
torchvision
torchaudio
transformers
opencv-python
fastapi
uvicorn
streamlit
pillow
pytorch-grad-cam
python-multipart
timm
numpy